ISSUE #2757 - Move tests to test folder
