+++ frontend/src/v4/modules/sagas.ts
@@ -16,8 +16,8 @@
  */
 
 import { all, fork } from 'redux-saga/effects';
-import projectsSaga from '../../v5/store/projects/projects.sagas';
-import teamspaces2Saga from '../../v5/store/teamspaces/teamspaces.sagas';
+import projectsSaga from '@/v5/store/projects/projects.sagas';
+import teamspaces2Saga from '@/v5/store/teamspaces/teamspaces.sagas';
 import activitiesSaga from './activities/activities.sagas';
 import authSaga from './auth/auth.sagas';
 import billingSaga from './billing/billing.sagas';
